[ETFE-1568] Minor refactor to models
